requirements.txt
# requirements.txt
# Core dependencies for MCP-enhanced CoT Spatiotemporal Behavior Generation

# Data processing and scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning and AI
torch>=1.9.0
transformers>=4.15.0

# Geographic and spatial processing
geopy>=2.2.0

# Time and date processing
pytz>=2021.3

# Configuration and data formats
PyYAML>=6.0
json5>=0.9.0

# System monitoring and parallel processing
psutil>=5.8.0
multiprocessing-logging>=0.3.0

# Statistical analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Development and testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.9.0
flake8>=4.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Performance optimization
numba>=0.56.0
cython>=0.29.0

# Memory and caching
joblib>=1.1.0
lru-dict>=1.1.0

# Logging and monitoring
structlog>=21.2.0
prometheus-client>=0.12.0

# API and web framework (for future extensions)
fastapi>=0.70.0
uvicorn>=0.15.0

# Database connectivity (optional)
sqlalchemy>=1.4.0
alembic>=1.7.0

# Jupyter notebook support (for analysis)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Export and serialization
openpyxl>=3.0.0
xlsxwriter>=3.0.0